% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{earthquakesdata}
\alias{earthquakesdata}
\title{NOAA Significant Earthquakes
 The dataset was obtained from the U.S. National Oceanographic and
 Atmospheric Administration (NOAA) on significant earthquakes around the world.
 This dataset contains information about 5,933 earthquakes over
 an approximately 4,000 year time span.}
\format{
An object of class \code{data.frame} with 6270 rows and 39 columns.
}
\source{
National Geophysical Data Center
}
\usage{
earthquakesdata
}
\description{
NOAA Significant Earthquakes
 The dataset was obtained from the U.S. National Oceanographic and
 Atmospheric Administration (NOAA) on significant earthquakes around the world.
 This dataset contains information about 5,933 earthquakes over
 an approximately 4,000 year time span.
}
\keyword{datasets}
