% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{The function Cleans the data for further analysis.}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{A Dataframe}
}
\value{
A Dataframe
}
\description{
The function Cleans the data for further analysis.
}
\details{
We would clean the data by

first creating a DATE column, which would be contain the Year,Month and Day
merged together in ymd format
The column `Search Parameter` contains no items, hence it would also be removed
The COUNTRY column would contain the country entry from the `Location Name` column
The 'Location.Name' would contain the city or region name from the `Location Name` column
}
\examples{
\dontrun{
data = system.file('extdata', 'earthquakesData.tsv', package ='masteringSoftwareDevelopmentinR')
data = read_data(data)
cleanedData <- eq_clean_data(data)
}
}
