% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_create_label.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Function to create a pop-up of text}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{A Dataframe}
}
\value{
A leaflet map with pop-up text containing location name, magnitude and total deaths
}
\description{
Function to create a pop-up of text
}
\examples{
\dontrun{
data = system.file('extdata', 'earthquakesData.tsv', package ='masteringSoftwareDevelopmentinR') \%>\%
data = read_data(data) \%>\%
eq_clean_data() \%>\%
 dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
 dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
 eq_map(annot_col = "popup_text")
 }

}
