% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{A Function to create map to visualize the earthquake Dataset}
\usage{
eq_map(annot_col, data)
}
\arguments{
\item{annot_col}{A charater}

\item{data}{A Dataframe}
}
\value{
A leaflet Map
}
\description{
A Function to create map to visualize the earthquake Dataset
}
\examples{
\dontrun{
data = system.file('extdata', 'earthquakesData.tsv', package ='masteringSoftwareDevelopmentinR')

eq_map(eq_clean_data(read_data(data)))
}

}
