% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_read_data.R
\name{read_data}
\alias{read_data}
\title{The function reads the data}
\usage{
read_data(filename)
}
\arguments{
\item{filename}{A character}
}
\value{
the file as a dataframe
}
\description{
The function reads the data
}
\examples{
\dontrun{
data = system.file('extdata', 'earthquakesData.tsv', package ='masteringSoftwareDevelopmentinR')
data = read_data(data)
}

}
